seperated types of pieces in different files
