+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(stocklion)
 
 set(CMAKE_CXX_STANDARD 11)
 
-add_executable(stocklion main.cpp datastructures/Piece.cpp datastructures/Piece.h datastructures/Pieces.cpp datastructures/Pieces.h)
+add_executable(stocklion main.cpp datastructures/Piece.cpp datastructures/Piece.h datastructures/Pawn.cpp datastructures/Pawn.h datastructures/Knight.cpp datastructures/Knight.h datastructures/Rook.cpp datastructures/Rook.h datastructures/Bishop.cpp datastructures/Bishop.h datastructures/King.cpp datastructures/King.h datastructures/Queen.cpp datastructures/Queen.h)
