+++ datastructures/Knight.cpp
@@ -1,0 +1,8 @@
+#include "Knight.h"
+
+Knight::Knight(const Position &initialPosition, const Color &initialPlayer) : Piece(initialPosition, initialPlayer) {}
+
+std::vector<Position> Knight::possibleMoves() {
+    std::vector<Position> positions;
+    return positions;
+}
