+++ datastructures/Knight.h
@@ -1,0 +1,13 @@
+#ifndef STOCKLION_KNIGHT_H
+#define STOCKLION_KNIGHT_H
+
+#include "Piece.h"
+
+class Knight : public Piece {
+public:
+    explicit Knight(const Position &initialPosition, const Color &initialPlayer);
+    std::vector<Position> possibleMoves() override;
+};
+
+
+#endif
